requirements.txt
streamlit
reportlab
qrcode
pillow
pdf2image
poppler-utils